renaming time to timing, profiling
