+++ amen/__init__.py
@@ -2,7 +2,7 @@
 # -*- coding: utf-8 -*-
 
 from .feature import Feature
-from .time import TimeSlice
+from .timing import TimeSlice
 from .version import version as __version__
 
 __all__ = ['audio', 'Feature', 'utils']
